add custom dataset

+++ mmaction/datasets/multilable_video_dataset.py
@@ -110,12 +110,12 @@ class MultilabelVideoDataset(BaseDataset):
             dict: Evaluation results dict.
         """
         results = torch.as_tensor(np.array(results), dtype=torch.float)
-        gt_labels = torch.as_tensor(np.array([ann['label'] for ann in self.video_infos]), dtype=torch.float)
+        gt_labels = torch.as_tensor(np.array([ann['label'] for ann in self.video_infos]), dtype=torch.long)
 
         results_sigmoid = results.sigmoid()
 
         loss = self.loss_fn(results, gt_labels)
-        cm = self.confusion_matrix(results_sigmoid, gt_labels.long())
+        cm = self.confusion_matrix(results_sigmoid, gt_labels)
         accuracy = self.accuracy(results_sigmoid, gt_labels)
         f1_score = self.f1_score(results_sigmoid, gt_labels)
         precision = self.prec(results_sigmoid, gt_labels)
